Better handling of same-velocity vertices at collapsed edges.
Handle case where event processing creates a wavefront edge that has two
vertices that move at the same velocity and that are in the same place.

Previously, we would characterize such an edge event as collapsing
always, which translated to a constraint collapse of time zero.
This is not ok when the situation happens during propagation, as our
event queue rightly refuses to go back in time.   Now edge collapse
times of always get turned into triangle collapse times of now.