Halfedge: rename is_input to is_emanating_input
Because that is what we used it for.  A halfedge has is_emanating_input
set if it is the source of a wavefront propagation (an input edge may
not be if we do not compute the straight skeleton in that component).

Also, add an is_input() method that returns true if either one of a
halfedge pair is an emanating input.