Fix accept_collapse_bounded_constrained_1 border case.
accept_collapse_bounded_constrained_1(): if the derivative of the signed
area at the collapse time is zero, i.e., this triangle collapses only
once and the area is a parabola with its vertex on the x-axis, then
whether or not we need to handle this event depends on whether this
is the time of a constraint collapse or not.  If the constraint
collapses, we do need to handle this.

Test inputs: degree-one-on-ch.graphml and srpg_octo0000121.graphml