Remove #if 0'd code
Remove some already disabled code in handling split events.

In a split event, a vertex v moved into an edge e.  Let e0 and e1 be the
edges incident to v.  Then the code would assert that none of e0 or e1
are parallel/collinear to e.  Since we can handle that case for a while
now, that checking code had been disabled and has now been removed.

If enabled, this would have triggered on, for instance,
test-data/weighted/srpg_octo0000035-random-3.graphml.

I have now also added a dedicated test case as
test-data/special-cases/collinear-split-event.graphml
which also triggeres the assertion, when enabled.  (Which it isn't now,
because this commits removes it.)