Make our CORE_ONE and CORE_ZERO variables references to the versions provided by CGAL
These variables are constant, static instances of the number type with values
of 1 and 0.  Up until now, they were their own variables, but now they have
become references to the version provided from CGAL.

This is probably a good idea in general, but it also fixes an actual issue.
If if we link against the surfer lib and then don't use it, CGAL
blows up at the end of the program with

} N4CORE14ConstDoubleRepE
} terminate called after throwing an instance of 'CGAL::Assertion_exception'
}   what():  CGAL ERROR: assertion violation!
} Expr: ! blocks.empty()
} File: /usr/include/CGAL/CORE/MemoryPool.h
} Line: 146

Now that CORE_{ONE,ZERO} are references, this no longer happens.