Make our own SURF_NDEBUG
This we define in cmake, similar to the normal NDEBUG define.  In
release builds, we define it, in debug builds its not defined.

The idea is that we can put SURF_NDEBUG into surfer's config.h.
Then we can include surfer's header in a different application and
since surfer's headers include surfer's config.h, they get the
proper ndebug/debug configuration that matches the compiled library.